Restructuring the agents endpoints in the Api. Splitting file per endpoint
